Split combinators by constructor type
