+++ mtpylon/service_schema/constructors/future_salt.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+from dataclasses import dataclass
+
+from mtpylon.utils import long
+
+
+@dataclass
+class FutureSalt:
+    valid_since: int
+    valid_until: int
+    salt: long
+
+    class Meta:
+        name = 'future_salt'
+        order = ('valid_since', 'valid_until', 'salt')
