+++ mtpylon/service_schema/constructors/msgs_state_req.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+from dataclasses import dataclass
+from typing import List
+
+from mtpylon.utils import long
+
+
+@dataclass
+class MsgsStateReq:
+    msg_ids: List[long]
+
+    class Meta:
+        name = 'msgs_state_req'
+        order = ('msg_ids', )
