+++ mtpylon/service_schema/constructors/pong.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+from dataclasses import dataclass
+
+from mtpylon.utils import long
+
+
+@dataclass
+class Pong:
+    msg_id: long
+    ping_id: long
+
+    class Meta:
+        name = 'pong'
+        order = ('msg_id', 'ping_id')
