+++ mtpylon/service_schema/constructors/rpc_error.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+from dataclasses import dataclass
+
+
+@dataclass
+class RpcError:
+    error_code: int
+    error_message: str
+
+    class Meta:
+        name = 'rpc_error'
+        order = ('error_code', 'error_message')
