+++ mtpylon/service_schema/constructors/server_dh_inner_data.py
@@ -1,0 +1,25 @@
+# -*- coding: utf-8 -*-
+from dataclasses import dataclass
+
+from mtpylon.utils import int128
+
+
+@dataclass
+class Server_DH_inner_data:
+    nonce: int128
+    server_nonce: int128
+    g: int
+    dh_prime: bytes
+    g_a: bytes
+    server_time: int
+
+    class Meta:
+        name = 'server_DH_inner_data'
+        order = (
+            'nonce',
+            'server_nonce',
+            'g',
+            'dh_prime',
+            'g_a',
+            'server_time'
+        )
